Move AccessReceivedMesh.
